requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic
pydantic-settings
email-validator
pandas
twilio
python-multipart
celery[redis]
phonenumbers
pytest
pytest-env
httpx
reportlab
xlsxwriter
